+++ tests/test_firefish.py
@@ -1,10 +1,19 @@
 from pyxavi.firefish import Firefish
+from pyxavi.config import Config
+from pyxavi.logger import Logger
+import logging
 from unittest.mock import Mock, patch, mock_open, call, MagicMock
 from unittest import TestCase
 import pytest
 import builtins
 import requests
 import json
+
+CONFIG = {"logger": {"name": "logger_test"}}
+
+
+def patch_config_read_file(self):
+    self._content = CONFIG
 
 
 @pytest.mark.parametrize(
@@ -174,6 +183,35 @@ def test_log_in_with_password():
         )
 
         assert instance.bearer_token == token
+
+
+@patch.object(Config, "read_file", new=patch_config_read_file)
+def test_init_with_logger():
+    client_id_filename = "client.secret"
+    client_id_client_name = "Client"
+    client_id_api_base_url = "https://social.devnamic.com"
+    client_id_content = "\n".join([client_id_api_base_url, client_id_client_name])
+
+    logger = Logger(config=Config()).get_logger()
+
+    with patch.object(builtins, "open", mock_open(read_data=client_id_content)):
+        instance = Firefish(client_id=client_id_filename, logger=logger)
+
+    assert isinstance(instance._logger, logging.Logger)
+    assert instance._logger.name == CONFIG["logger"]["name"]
+
+
+def test_init_without_logger():
+    client_id_filename = "client.secret"
+    client_id_client_name = "Client"
+    client_id_api_base_url = "https://social.devnamic.com"
+    client_id_content = "\n".join([client_id_api_base_url, client_id_client_name])
+
+    with patch.object(builtins, "open", mock_open(read_data=client_id_content)):
+        instance = Firefish(client_id=client_id_filename)
+
+    assert isinstance(instance._logger, logging.Logger)
+    assert instance._logger.name == "firefish_wrapper"
 
 
 @pytest.mark.parametrize(