+++ tests/test_mastodon_connection_params.py
@@ -1,6 +1,6 @@
 from pyxavi.mastodon_helper import\
     MastodonConnectionParams, MastodonCredentials, MastodonUser, MastodonStatusParams,\
-    StatusPostContentType, StatusPostVisibility
+    StatusPostContentType, StatusPostVisibility, MastodonInstanceType
 from unittest import TestCase
 import pytest
 
@@ -15,7 +15,7 @@ def test_to_dict_minimal():
     d = MastodonConnectionParams().to_dict()
 
     assert d["app_name"] is None
-    assert d["instance_type"] == MastodonConnectionParams.TYPE_MASTODON
+    assert d["instance_type"] == MastodonInstanceType.MASTODON
     assert d["api_base_url"] is None
     assert d["credentials"] is None
     assert isinstance(d["status_params"], dict)
@@ -29,7 +29,7 @@ def test_from_dict_minimal():
     instance = MastodonConnectionParams.from_dict({})
 
     assert instance.app_name is None
-    assert instance.instance_type == MastodonConnectionParams.TYPE_MASTODON
+    assert instance.instance_type == MastodonInstanceType.MASTODON
     assert instance.api_base_url is None
     assert instance.credentials is None
     assert isinstance(instance.status_params, MastodonStatusParams)
@@ -47,15 +47,15 @@ def test_instantiate_with_app_name():
 
 
 def test_instantiate_with_instance_type():
-    instance = MastodonConnectionParams(instance_type=MastodonConnectionParams.TYPE_FIREFISH)
+    instance = MastodonConnectionParams(instance_type=MastodonInstanceType.FIREFISH)
 
-    assert instance.instance_type == MastodonConnectionParams.TYPE_FIREFISH
+    assert instance.instance_type == MastodonInstanceType.FIREFISH
 
 
 def test_instantiate_without_instace_type_defers_to_mastodon():
     instance = MastodonConnectionParams()
 
-    assert instance.instance_type == MastodonConnectionParams.TYPE_MASTODON
+    assert instance.instance_type == MastodonInstanceType.MASTODON
 
 
 def test_instantiate_with_api_base_url():